modified where web app gets data and delete unused files, also a renaming
