--- test/fake-mongo-server.js
@@ -1,6 +1,0 @@
-var mongoose = require('mongoose')
-  , mongodbFs = require('mongodb-fs');
-
-exports.initialize = function () {
-
-}
