--- test/lineSaver-spec.js
@@ -1,4 +1,0 @@
-var test = require('tape');
-var lineSaver = require('../lib/save-handlers/lineSaver');
-
-
